models/anomaly/requirements.txt
about-time==4.2.1
absl-py==2.3.1
aiofiles==24.1.0
aiohappyeyeballs==2.6.1
aiohttp==3.12.15
aiosignal==1.4.0
albucore==0.0.24
alembic==1.16.4
alive-progress==3.3.0
annotated-types==0.7.0
anomalib==2.1.0
antlr4-python3-runtime==4.9.3
anyio==4.10.0
argcomplete==3.6.2
argon2-cffi==25.1.0
argon2-cffi-bindings==25.1.0
arrow==1.3.0
asttokens==3.0.0
async-lru==2.0.5
attrs==25.3.0
autograd==1.8.0
av==15.0.0
babel==2.17.0
bcrypt==4.3.0
beautifulsoup4==4.13.4
bleach==6.2.0
blinker==1.9.0
boto3==1.40.10
botocore==1.40.10
Brotli==1.1.0
cachetools==5.5.2
certifi==2025.8.3
cffi==1.17.1
charset-normalizer==3.4.3
choreographer==1.0.9
click==8.2.1
cloudpickle==3.1.1
cma==4.3.0
colorama==0.4.6
comet_ml==3.50.0
comm==0.2.3
configobj==5.0.9
contourpy==1.3.3
cryptography==45.0.6
cycler==0.12.1
dacite==1.7.0
databricks-sdk==0.63.0
debugpy==1.8.16
decorator==5.2.1
defusedxml==0.7.1
Deprecated==1.2.18
dill==0.4.0
distro==1.9.0
dnspython==2.7.0
docker==7.1.0
docstring_parser==0.17.0
dotenv==0.9.9
dulwich==0.24.1
einops==0.8.1
everett==3.1.0
executing==2.2.0
fastapi==0.116.1
fastjsonschema==2.21.2
ffmpy==0.6.1
fiftyone==1.7.2
fiftyone-brain==0.21.2
fiftyone_db==1.3.0
filelock==3.18.0
Flask==3.1.1
fonttools==4.59.1
fqdn==1.5.1
FrEIA==0.2
frozenlist==1.7.0
fsspec==2025.7.0
ftfy==6.3.1
future==1.0.0
gitdb==4.0.12
GitPython==3.1.45
glob2==0.7
google-auth==2.40.3
gradio==5.42.0
gradio_client==1.11.1
graphemeu==0.7.2
graphene==3.4.3
graphql-core==3.2.6
graphql-relay==3.2.0
greenlet==3.2.4
groovy==0.1.2
grpcio==1.74.0
h11==0.16.0
h2==4.2.0
hpack==4.1.0
httpcore==1.0.9
httpx==0.28.1
huggingface-hub==0.34.4
humanize==4.12.3
Hypercorn==0.17.3
hyperframe==6.1.0
idna==3.10
imageio==2.37.0
importlib_metadata==8.7.0
importlib_resources==6.5.2
inflate64==1.0.3
ipykernel==6.30.1
ipython==9.4.0
ipython_pygments_lexers==1.1.1
ipywidgets==8.1.7
isoduration==20.11.0
itsdangerous==2.2.0
jedi==0.19.2
Jinja2==3.1.6
jiter==0.10.0
jmespath==1.0.1
joblib==1.5.1
json5==0.12.1
jsonargparse==4.40.2
jsonlines==4.0.0
jsonpointer==3.0.0
jsonschema==4.25.0
jsonschema-specifications==2025.4.1
jupyter-events==0.12.0
jupyter-lsp==2.2.6
jupyter_client==8.6.3
jupyter_core==5.8.1
jupyter_server==2.16.0
jupyter_server_terminals==0.5.3
jupyterlab==4.4.5
jupyterlab_pygments==0.3.0
jupyterlab_server==2.27.3
jupyterlab_widgets==3.0.15
kaleido==1.0.0
kiwisolver==1.4.9
kornia==0.8.1
kornia_rs==0.1.9
lark==1.2.2
lazy_loader==0.4
lia-web==0.2.3
lightning==2.5.3
lightning-utilities==0.15.2
logistro==1.1.0
Mako==1.3.10
Markdown==3.8.2
markdown-it-py==4.0.0
MarkupSafe==3.0.2
matplotlib==3.10.5
matplotlib-inline==0.1.7
mdurl==0.1.2
mistune==3.1.3
mlflow==3.2.0
mlflow-skinny==3.2.0
mlflow-tracing==3.2.0
mongoengine==0.29.1
motor==3.6.1
mpmath==1.3.0
multidict==6.6.4
multivolumefile==0.2.3
narwhals==2.1.1
natsort==8.4.0
nbclient==0.10.2
nbconvert==7.16.6
nbformat==5.10.4
nest-asyncio==1.6.0
networkx==3.4.2
ninja==1.11.1.4
nncf==2.17.0
notebook==7.4.5
notebook_shim==0.2.4
numpy==2.2.6
ollama==0.5.3
omegaconf==2.3.0
onnx==1.18.0
open-clip-torch==2.24.0
openai==1.99.9
opencv-python==4.12.0.88
opencv-python-headless==4.12.0.88
opentelemetry-api==1.36.0
opentelemetry-sdk==1.36.0
opentelemetry-semantic-conventions==0.57b0
openvino==2025.2.0
openvino-telemetry==2025.2.0
orjson==3.11.2
overrides==7.7.0
packaging==25.0
pandas==2.2.3
pandocfilters==1.5.1
paramiko==3.5.1
parso==0.8.4
pillow==11.3.0
platformdirs==4.3.8
plotly==6.3.0
pprintpp==0.4.0
precision_lens @ file:///C:/Github/MLOps-Precision-Lens/shared_libraries
priority==2.0.0
prometheus_client==0.22.1
prompt_toolkit==3.0.51
propcache==0.3.2
protobuf==6.32.0
psutil==7.0.0
pure_eval==0.2.3
py7zr==1.0.0
pyarrow==21.0.0
pyasn1==0.6.1
pyasn1_modules==0.4.2
pybcj==1.0.6
pycparser==2.22
pycryptodomex==3.23.0
pydantic==2.11.7
pydantic_core==2.33.2
pydash==8.0.5
pydot==3.0.4
pydub==0.25.1
Pygments==2.19.2
pymongo==4.9.2
pymoo==0.6.1.5
PyNaCl==1.5.0
pyparsing==3.2.3
pyppmd==1.2.0
python-box==6.1.0
python-dateutil==2.9.0.post0
python-dotenv==1.1.1
python-json-logger==3.3.0
python-multipart==0.0.20
pytorch-lightning==2.5.3
pytz==2025.2
pywin32==311
pywinpty==3.0.0
PyYAML==6.0.2
pyzmq==27.0.1
pyzstd==0.17.0
rarfile==4.2
referencing==0.36.2
regex==2025.7.34
requests==2.32.4
requests-toolbelt==1.0.0
retrying==1.4.2
rfc3339-validator==0.1.4
rfc3986-validator==0.1.1
rfc3987-syntax==1.1.0
rich==14.1.0
rich-argparse==1.7.1
rpds-py==0.27.0
rsa==4.9.1
rtree==1.4.1
ruff==0.12.9
s3transfer==0.13.1
safehttpx==0.1.6
safetensors==0.6.2
scikit-image==0.25.2
scikit-learn==1.7.1
scipy==1.16.1
semantic-version==2.10.0
Send2Trash==1.8.3
sentencepiece==0.2.1
sentry-sdk==2.35.0
setuptools==80.9.0
shellingham==1.5.4
simplejson==3.20.1
simsimd==6.5.1
six==1.17.0
smmap==5.0.2
sniffio==1.3.1
sortedcontainers==2.4.0
soupsieve==2.7
SQLAlchemy==2.0.43
sqlparse==0.5.3
sse-starlette==0.10.3
sseclient-py==1.8.0
stack-data==0.6.3
starlette==0.47.2
strawberry-graphql==0.278.1
stringzilla==3.12.6
sympy==1.14.0
tabulate==0.9.0
tensorboard==2.20.0
tensorboard-data-server==0.7.2
terminado==0.18.1
texttable==1.7.0
threadpoolctl==3.6.0
tifffile==2025.6.11
timm==1.0.19
tinycss2==1.4.0
tokenizers==0.21.4
tomlkit==0.13.3
torch==2.8.0+cu128
torchmetrics==1.8.1
torchvision==0.23.0
tornado==6.5.2
tqdm==4.67.1
traitlets==5.14.3
transformers==4.55.2
typer==0.16.0
types-python-dateutil==2.9.0.20250809
typeshed_client==2.8.2
typing-inspection==0.4.1
typing_extensions==4.14.1
tzdata==2025.2
tzlocal==5.3.1
universal-analytics-python3==1.1.1
uri-template==1.3.0
urllib3==2.5.0
uvicorn==0.35.0
voxel51-eta==0.14.4
waitress==3.0.2
wandb==0.21.1
wcwidth==0.2.13
webcolors==24.11.1
webencodings==0.5.1
websocket-client==1.8.0
websockets==15.0.1
Werkzeug==3.1.3
widgetsnbextension==4.0.14
wrapt==1.17.3
wsproto==1.2.0
wurlitzer==3.1.1
xmltodict==0.14.2
yarl==1.20.1
zipp==3.23.0
